strengthen tests from audit (prio 2-5)
Prio 2 – missing assertions: resize_reject_read_only actually calls
resize(), write_spanning_two_clusters verifies data readback,
read_at_offset_within_cluster checks before-buffer, metadata error
display asserts size field, repair test checks total_errors.

Prio 3 – misleading names/logic: reject_l2_table_offset renamed to
_beyond_eof with specific error match, find_snapshot_by_name exercises
actual find via delete+apply, compress round-trip uses compress_cluster.

Prio 4 – inconsistent setup: write_in_place sets refcount for data
cluster, custom_allocator uses SequentialAllocator with second alloc,
max_refcount_64bit adds refcount_order assertion.

Prio 5 – fragile patterns: distinctive multi-digit test values in
error.rs, TestImage::wrap() constructor, byte-identical round-trip
check, full-cluster verification in stress test.

+++ src/format/header.rs
@@ -464,7 +464,7 @@ impl Header {
         // Only deflate compression is supported
         if self.compression_type != COMPRESSION_DEFLATE {
             return Err(Error::UnsupportedCompressionType {
-                compression_type: self.compression_type as u8,
+                compression_type: self.compression_type,
             });
         }
